zoho/backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
requests
python-dotenv
schedule
gunicorn
google-generativeai
psycopg2-binary
